Cache configuration directory state to avoid costly nix calls
Running nix-instantiate or nix-build to query the configuration always
takes several seconds. To avoid these delays, we now cache the state of
all git-files inside the configuration directory. If at least one of the
files changes, we run nix and cache the state. If no file has changed,
we can reuse previously built artefacts, such as confctl settings and
machine list.

For this reason, confctl settings are now also built using nix-build, so
that we can cache the values.